[ez] Move global components to a separate folder
I think this is cleaner and easier to manage. I thought the name "config" for the folder was too overloaded, so think "global" is better

## Test Plan
No functional changes
