+++ python/src/aiconfig/editor/client/src/components/global/ConfigNameDescription.tsx
